requirement.txt
webdriver-manager
colorma
pytubefix
selenium
ffmpeg-python

